build.sh and Dockerfile updated

+++ make.sh
@@ -2,8 +2,10 @@
 set -x 
 
 TGZ=debian.tgz
-sudo debootstrap --variant=minbase wheezy wheezy
-sudo tar -cz  -C wheezy --numeric-owner . > $TGZ
-sudo rm -rf wheezy
+if [ ! -f $TGZ ]; then
+    sudo debootstrap --variant=minbase wheezy wheezy
+    sudo tar -cz  -C wheezy --numeric-owner . > $TGZ
+    sudo rm -rf wheezy
+fi
 
 docker build -t deliverous/wheezy .
